improve mobile and update readme
