+++ site/src/App.tsx
@@ -14,7 +14,7 @@ function PreviewFallback() {
   return (
     <section id="preview" className="py-24 px-6">
       <div className="max-w-6xl mx-auto">
-        <div className="h-[600px] bg-slate-900/40 border border-slate-800/50 rounded-sm flex items-center justify-center">
+        <div className="h-150 bg-slate-900/40 border border-slate-800/50 rounded-sm flex items-center justify-center">
           <div className="text-slate-500 text-sm uppercase tracking-wide font-bold">
             Loading preview...
           </div>
@@ -31,7 +31,7 @@ export function App() {
         href="https://github.com/tariknz/irdashies"
         target="_blank"
         rel="noopener noreferrer"
-        className="fixed top-4 right-4 z-50 flex items-center gap-2 px-3 py-2 bg-slate-800/80 hover:bg-slate-700/80 backdrop-blur-sm border border-slate-700/50 rounded-sm text-slate-300 hover:text-white text-sm font-medium transition-colors"
+        className="fixed top-4 right-4 z-50 hidden sm:flex items-center gap-2 px-3 py-2 bg-slate-800/80 hover:bg-slate-700/80 backdrop-blur-sm border border-slate-700/50 rounded-sm text-slate-300 hover:text-white text-sm font-medium transition-colors"
       >
         <GithubLogo size={20} weight="fill" />
         GitHub
