+++ site/src/sections/LivePreview.tsx
@@ -3,12 +3,25 @@ import {
   useCallback,
   useEffect,
   useRef,
-  memo,
-  type ReactNode,
-  type CSSProperties,
+  useSyncExternalStore,
 } from 'react';
+
+const MOBILE_QUERY = '(max-width: 639px)';
+const mobileMedia =
+  typeof window !== 'undefined' ? window.matchMedia(MOBILE_QUERY) : null;
+
+function subscribeMobile(cb: () => void) {
+  mobileMedia?.addEventListener('change', cb);
+  return () => mobileMedia?.removeEventListener('change', cb);
+}
+function getSnapshotMobile() {
+  return mobileMedia?.matches ?? false;
+}
+
+function useIsMobile() {
+  return useSyncExternalStore(subscribeMobile, getSnapshotMobile);
+}
 import { LivePreviewProvider } from '../utils/mockSetup';
-import { useGeneralSettings } from '../../../src/frontend/context/DashboardContext/DashboardContext';
 import { WidgetErrorBoundary } from '../components/WidgetErrorBoundary';
 import { DashboardReady } from '../components/DashboardReady';
 import { Standings } from '../../../src/frontend/components/Standings/Standings';
@@ -27,28 +40,15 @@ import { RejoinIndicator } from '../../../src/frontend/components/RejoinIndicato
 import { PitlaneHelper } from '../../../src/frontend/components/PitlaneHelper/PitlaneHelper';
 import { LapTimeLog } from '../../../src/frontend/components/LapTimeLog/LapTimeLog';
 import { SlowCarAhead } from '../../../src/frontend/components/SlowCarAhead/SlowCarAhead';
-import {
-  useDragWidget,
-  useResizeWidget,
-  ResizeHandles,
-} from '../../../src/frontend/components/WidgetContainer';
-import {
-  ArrowsOutCardinal,
-  ArrowUp,
-  CursorClick,
-  GearSix,
-  X,
-} from '@phosphor-icons/react';
 import { PreviewSettingsButton } from '../components/PreviewSettingsPanel';
 import { defaultDashboard } from '../../../src/types/defaultDashboard';
-import { useDashboard } from '../../../src/frontend/context/DashboardContext/DashboardContext';
-
-interface WidgetPosition {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
+import { CoachMarks } from './CoachMarks';
+import {
+  PreviewWidgetItem,
+  WidgetFrame,
+  type WidgetPosition,
+} from './PreviewWidgetItem';
+import { ActiveWidgetSync } from './ActiveWidgetSync';
 
 /** Look up a widget's default layout size from the project's defaultDashboard. */
 function getDefaultSize(widgetId: string): { width: number; height: number } {
@@ -168,263 +168,8 @@ function buildDefaultPositions(
   return positions;
 }
 
-/**
- * Wraps a widget in the overlay-window theme container.
- * Reads theme settings from the dashboard's generalSettings via useGeneralSettings,
- * mirroring the real ThemeManager behaviour.
- */
-function WidgetFrame({ children }: { children: ReactNode }) {
-  const generalSettings = useGeneralSettings();
-  const fontSize = generalSettings?.fontSize ?? 'sm';
-  const colorPalette = generalSettings?.colorPalette ?? 'default';
-  const fontType = generalSettings?.fontType ?? 'lato';
-  const fontWeight = generalSettings?.fontWeight ?? 'normal';
-
-  return (
-    <div
-      className={[
-        'overlay-window',
-        `overlay-theme-${fontSize}`,
-        `overlay-theme-color-${colorPalette}`,
-        `overlay-theme-font-face-${fontType}`,
-        `overlay-theme-font-weight-${fontWeight}`,
-        'w-full h-full overflow-hidden',
-      ].join(' ')}
-    >
-      {children}
-    </div>
-  );
-}
-
-/**
- * A single draggable/resizable widget in the preview canvas.
- */
-const PreviewWidgetItem = memo(function PreviewWidgetItem({
-  widgetId,
-  label,
-  component: Component,
-  position,
-  isSelected,
-  onPositionChange,
-  onSelect,
-  onDeselect,
-  onDoubleClick,
-}: {
-  widgetId: string;
-  label: string;
-  component: React.ComponentType;
-  position: WidgetPosition;
-  isSelected: boolean;
-  onPositionChange: (id: string, pos: WidgetPosition) => void;
-  onSelect: (id: string) => void;
-  onDeselect: () => void;
-  onDoubleClick: (id: string) => void;
-}) {
-  const [localLayout, setLocalLayout] = useState(position);
-  const debounceRef = useRef<ReturnType<typeof setTimeout> | undefined>(
-    undefined
-  );
-
-  useEffect(() => {
-    setLocalLayout(position);
-  }, [position]);
-
-  const handleLayoutChange = useCallback(
-    (newLayout: WidgetPosition) => {
-      setLocalLayout(newLayout);
-      if (debounceRef.current) clearTimeout(debounceRef.current);
-      debounceRef.current = setTimeout(() => {
-        onPositionChange(widgetId, newLayout);
-      }, 100);
-    },
-    [widgetId, onPositionChange]
-  );
-
-  const flushPendingSave = useCallback(() => {
-    if (debounceRef.current) {
-      clearTimeout(debounceRef.current);
-      debounceRef.current = undefined;
-    }
-    onPositionChange(widgetId, localLayout);
-  }, [widgetId, localLayout, onPositionChange]);
-
-  const { isDragging, dragHandleProps } = useDragWidget({
-    layout: localLayout,
-    onLayoutChange: handleLayoutChange,
-    enabled: true,
-  });
-
-  const { isResizing, getResizeHandleProps } = useResizeWidget({
-    layout: localLayout,
-    onLayoutChange: handleLayoutChange,
-    enabled: true,
-  });
-
-  const isInteracting = isDragging || isResizing;
-  const prevInteractingRef = useRef(isInteracting);
-
-  useEffect(() => {
-    const wasInteracting = prevInteractingRef.current;
-    prevInteractingRef.current = isInteracting;
-    if (wasInteracting && !isInteracting) {
-      flushPendingSave();
-    }
-  }, [isInteracting, flushPendingSave]);
-
-  useEffect(() => {
-    return () => {
-      if (debounceRef.current) clearTimeout(debounceRef.current);
-    };
-  }, []);
-
-  const containerStyle: CSSProperties = {
-    position: 'absolute',
-    left: localLayout.x,
-    top: localLayout.y,
-    width: localLayout.width,
-    height: localLayout.height,
-  };
-
-  return (
-    <div style={containerStyle} data-widget-id={widgetId}>
-      <div
-        {...dragHandleProps}
-        className="w-full h-full overflow-hidden text-white relative"
-        onClick={() => onSelect(widgetId)}
-        onDoubleClick={(e) => {
-          e.stopPropagation();
-          onDoubleClick(widgetId);
-        }}
-      >
-        {/* Selection border */}
-        {isSelected && (
-          <div className="absolute inset-0 border-dashed border-2 border-sky-500 pointer-events-none z-20 flex items-start justify-end p-2">
-            <div className="flex items-center gap-2 bg-sky-500 text-white text-sm font-semibold px-2 py-1 rounded pointer-events-auto">
-              <ArrowsOutCardinal size={14} />
-              <span>{label}</span>
-              <button
-                onClick={(e) => {
-                  e.stopPropagation();
-                  onDeselect();
-                }}
-                className="ml-1 hover:bg-sky-600 rounded p-0.5 transition-colors"
-                title="Close"
-              >
-                <X size={14} />
-              </button>
-            </div>
-          </div>
-        )}
-        {/* Widget content */}
-        <WidgetFrame>
-          <WidgetErrorBoundary widgetName={label}>
-            <Component />
-          </WidgetErrorBoundary>
-        </WidgetFrame>
-      </div>
-
-      <ResizeHandles getResizeHandleProps={getResizeHandleProps} />
-    </div>
-  );
-});
-
-/**
- * Syncs the toolbar's activeWidgets set into the dashboard context so the
- * settings panel's enabled toggle stays in sync with the Frame toolbar.
- */
-function ActiveWidgetSync({ activeWidgets }: { activeWidgets: Set<string> }) {
-  const { currentDashboard, onDashboardUpdated } = useDashboard();
-  const prevActiveRef = useRef<Set<string>>(activeWidgets);
-
-  useEffect(() => {
-    if (!currentDashboard || !onDashboardUpdated) return;
-    // Only sync when activeWidgets actually changed (skip initial render)
-    if (prevActiveRef.current === activeWidgets) return;
-    prevActiveRef.current = activeWidgets;
-
-    const needsUpdate = currentDashboard.widgets.some((w) => {
-      const id = w.type ?? w.id;
-      return w.enabled !== activeWidgets.has(id);
-    });
-
-    if (!needsUpdate) return;
-
-    const updatedWidgets = currentDashboard.widgets.map((w) => {
-      const id = w.type ?? w.id;
-      const shouldBeEnabled = activeWidgets.has(id);
-      if (w.enabled === shouldBeEnabled) return w;
-      return { ...w, enabled: shouldBeEnabled };
-    });
-
-    onDashboardUpdated({ ...currentDashboard, widgets: updatedWidgets });
-  }, [activeWidgets, currentDashboard, onDashboardUpdated]);
-
-  return null;
-}
-
-/**
- * First-visit coach marks overlay — three callouts pointing at the toolbar,
- * a widget, and the configure button. Dismissed on first click anywhere.
- */
-function CoachMarks({ onDismiss }: { onDismiss: () => void }) {
-  const [fading, setFading] = useState(false);
-
-  const dismiss = useCallback(() => {
-    setFading(true);
-    setTimeout(onDismiss, 300);
-  }, [onDismiss]);
-
-  useEffect(() => {
-    const handler = () => dismiss();
-    window.addEventListener('pointerdown', handler, { once: true });
-    return () => window.removeEventListener('pointerdown', handler);
-  }, [dismiss]);
-
-  return (
-    <div
-      className={[
-        'absolute inset-0 z-50 pointer-events-none transition-opacity duration-300',
-        fading ? 'opacity-0' : 'opacity-100',
-      ].join(' ')}
-    >
-      {/* Scrim */}
-      <div className="absolute inset-0 bg-black/30" />
-
-      {/* Callout: toolbar toggles — top-left, below toolbar */}
-      <div className="absolute top-14 left-16 flex flex-col items-start gap-2 animate-pulse">
-        <ArrowUp size={24} className="text-sky-400 ml-5" />
-        <div className="bg-sky-500 text-white text-sm font-semibold px-4 py-2 rounded-md shadow-lg flex items-center gap-2">
-          <CursorClick size={18} />
-          Click to enable / disable widgets
-        </div>
-      </div>
-
-      {/* Callout: configure button — top-right, below toolbar */}
-      <div className="absolute top-14 right-4 flex flex-col items-end gap-2 animate-pulse">
-        <ArrowUp size={24} className="text-sky-400 mr-5" />
-        <div className="bg-sky-500 text-white text-sm font-semibold px-4 py-2 rounded-md shadow-lg flex items-center gap-2">
-          <GearSix size={18} />
-          Customize widget settings
-        </div>
-      </div>
-
-      {/* Callout: widget interaction — center of canvas */}
-      <div className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 flex flex-col items-center gap-2 animate-pulse">
-        <div className="bg-sky-500 text-white text-sm font-semibold px-4 py-2 rounded-md shadow-lg flex items-center gap-2">
-          <ArrowsOutCardinal size={18} />
-          Drag widgets to move, double-click for settings
-        </div>
-      </div>
-
-      {/* Dismiss hint */}
-      <div className="absolute top-1/2 left-1/2 -translate-x-1/2 translate-y-12 bg-black/50 text-white/80 text-sm uppercase tracking-widest font-semibold px-4 py-1.5 rounded-md">
-        Click anywhere to dismiss
-      </div>
-    </div>
-  );
-}
-
 export function LivePreview() {
+  const isMobile = useIsMobile();
   const [activeWidgets, setActiveWidgets] = useState<Set<string>>(
     () => new Set(AVAILABLE_WIDGETS.filter((w) => w.defaultOn).map((w) => w.id))
   );
@@ -459,6 +204,10 @@ export function LivePreview() {
 
   const toggleWidget = (id: string) => {
     setActiveWidgets((prev) => {
+      if (isMobile) {
+        // Radio behaviour: one widget at a time
+        return prev.has(id) ? new Set<string>() : new Set([id]);
+      }
       const next = new Set(prev);
       if (next.has(id)) {
         next.delete(id);
@@ -500,9 +249,9 @@ export function LivePreview() {
           See It In <span className="text-red-600">Action</span>
         </h2>
         <p className="text-sm text-slate-400 mt-1">
-          Toggle widgets in the toolbar to enable them. Click a widget to select
-          it, drag to reposition, resize from edges, or double-click for
-          settings.
+          {isMobile
+            ? 'Tap the toolbar to preview different widgets. Try the full interactive experience on desktop.'
+            : 'Toggle widgets in the toolbar to enable them. Click a widget to select it, drag to reposition, resize from edges, or double-click for settings.'}
         </p>
       </div>
 
@@ -531,17 +280,19 @@ export function LivePreview() {
         {/* Preview frame */}
         <div className="mx-auto w-full max-w-450 flex-1 flex flex-col min-h-0">
           <div className="relative rounded-sm border border-slate-700/50 overflow-hidden carbon-fiber flex-1 flex flex-col">
-            {/* Frame toolbar — traffic lights + controls integrated */}
+            {/* Frame toolbar */}
             <div className="flex-none flex items-center gap-3 px-4 py-2 bg-slate-900/80 border-b border-slate-700/50">
-              {/* Traffic lights */}
-              <div className="flex gap-1.5 flex-none">
-                <div className="w-2.5 h-2.5 rounded-full bg-red-500/60" />
-                <div className="w-2.5 h-2.5 rounded-full bg-yellow-500/60" />
-                <div className="w-2.5 h-2.5 rounded-full bg-green-500/60" />
-              </div>
-
-              {/* Divider */}
-              <div className="w-px h-4 bg-slate-700/60 flex-none" />
+              {/* Traffic lights — desktop only */}
+              {!isMobile && (
+                <>
+                  <div className="flex gap-1.5 flex-none">
+                    <div className="w-2.5 h-2.5 rounded-full bg-red-500/60" />
+                    <div className="w-2.5 h-2.5 rounded-full bg-yellow-500/60" />
+                    <div className="w-2.5 h-2.5 rounded-full bg-green-500/60" />
+                  </div>
+                  <div className="w-px h-4 bg-slate-700/60 flex-none" />
+                </>
+              )}
 
               {/* Widget toggles — scrollable */}
               <div className="flex items-center gap-2 flex-1 min-w-0 overflow-x-auto scrollbar-hide">
@@ -550,7 +301,10 @@ export function LivePreview() {
                     key={widget.id}
                     onClick={() => toggleWidget(widget.id)}
                     className={[
-                      'px-2 py-0.5 text-[10px] font-bold uppercase tracking-wide rounded-sm transition-all border whitespace-nowrap flex-none',
+                      'font-bold uppercase tracking-wide rounded-sm transition-all border whitespace-nowrap flex-none',
+                      isMobile
+                        ? 'px-2.5 py-1 text-[11px]'
+                        : 'px-2 py-0.5 text-[10px]',
                       activeWidgets.has(widget.id)
                         ? 'border-red-600/50 bg-red-600/10 text-slate-200'
                         : 'border-slate-700/50 text-slate-400 hover:text-slate-200 hover:bg-slate-700/50 hover:border-slate-600/50',
@@ -561,30 +315,32 @@ export function LivePreview() {
                 ))}
               </div>
 
-              {/* Divider */}
-              <div className="w-px h-4 bg-slate-700/60 flex-none" />
-
-              {/* Settings */}
-              <div className="flex-none">
-                <PreviewSettingsButton
-                  activeWidgets={activeWidgets}
-                  onToggleWidget={toggleWidget}
-                  openToWidget={settingsOpenWidget}
-                  onClose={() => setSettingsOpenWidget(null)}
-                />
-              </div>
+              {/* Settings — desktop only */}
+              {!isMobile && (
+                <>
+                  <div className="w-px h-4 bg-slate-700/60 flex-none" />
+                  <div className="flex-none">
+                    <PreviewSettingsButton
+                      activeWidgets={activeWidgets}
+                      onToggleWidget={toggleWidget}
+                      openToWidget={settingsOpenWidget}
+                      onClose={() => setSettingsOpenWidget(null)}
+                    />
+                  </div>
+                </>
+              )}
             </div>
 
-            {/* Coach marks overlay */}
-            {showCoachMarks && (
+            {/* Coach marks overlay — desktop only */}
+            {!isMobile && showCoachMarks && (
               <CoachMarks onDismiss={() => setShowCoachMarks(false)} />
             )}
 
-            {/* Widget canvas — absolute positioned like the real DashboardView */}
+            {/* Widget canvas */}
             <div
               ref={canvasRef}
               className="relative flex-1 overflow-hidden"
-              onClick={() => setSelectedWidget(null)}
+              onClick={isMobile ? undefined : () => setSelectedWidget(null)}
             >
               {/* Background video */}
               <video
@@ -596,27 +352,47 @@ export function LivePreview() {
               >
                 <source src="/preview-bg.mp4" type="video/mp4" />
               </video>
+
               <DashboardReady>
-                {AVAILABLE_WIDGETS.filter((w) => activeWidgets.has(w.id)).map(
-                  (widget) => {
-                    const pos = positions[widget.id];
-                    if (!pos) return null;
-                    return (
-                      <PreviewWidgetItem
-                        key={widget.id}
-                        widgetId={widget.id}
-                        label={widget.label}
-                        component={widget.component}
-                        position={pos}
-                        isSelected={selectedWidget === widget.id}
-                        onPositionChange={handlePositionChange}
-                        onSelect={handleSelect}
-                        onDeselect={handleDeselect}
-                        onDoubleClick={handleDoubleClick}
-                      />
-                    );
-                  }
-                )}
+                {isMobile
+                  ? // Mobile: single static widget, no interactions
+                    (() => {
+                      const active = AVAILABLE_WIDGETS.find((w) =>
+                        activeWidgets.has(w.id)
+                      );
+                      if (!active) return null;
+                      const Component = active.component;
+                      return (
+                        <div className="relative w-full h-full">
+                          <WidgetFrame>
+                            <WidgetErrorBoundary widgetName={active.label}>
+                              <Component />
+                            </WidgetErrorBoundary>
+                          </WidgetFrame>
+                        </div>
+                      );
+                    })()
+                  : // Desktop: draggable/resizable widgets
+                    AVAILABLE_WIDGETS.filter((w) =>
+                      activeWidgets.has(w.id)
+                    ).map((widget) => {
+                      const pos = positions[widget.id];
+                      if (!pos) return null;
+                      return (
+                        <PreviewWidgetItem
+                          key={widget.id}
+                          widgetId={widget.id}
+                          label={widget.label}
+                          component={widget.component}
+                          position={pos}
+                          isSelected={selectedWidget === widget.id}
+                          onPositionChange={handlePositionChange}
+                          onSelect={handleSelect}
+                          onDeselect={handleDeselect}
+                          onDoubleClick={handleDoubleClick}
+                        />
+                      );
+                    })}
               </DashboardReady>
             </div>
           </div>
